refactor: :truck: Reorganiza arquivos. Move componentes de /ui para /custom
