+++ src/app/(app)/components/monitored-plate-form-dialog.tsx
@@ -5,6 +5,10 @@ import { useEffect, useState } from 'react'
 import { Controller, useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 
+import { InputError } from '@/components/custom/input-error'
+import MultipleSelector from '@/components/custom/multiselect-with-search'
+import { SelectWithSearch } from '@/components/custom/select-with-search'
+import { Spinner } from '@/components/custom/spinner'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
@@ -14,11 +18,7 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog'
 import { Input } from '@/components/ui/input'
-import { InputError } from '@/components/ui/input-error'
 import { Label } from '@/components/ui/label'
-import MultipleSelector from '@/components/ui/multiselect-with-search'
-import { SelectWithSearch } from '@/components/ui/select-with-search'
-import { Spinner } from '@/components/ui/spinner'
 import { Switch } from '@/components/ui/switch'
 import { Textarea } from '@/components/ui/textarea'
 import {
