+++ src/app/(app)/mapa/components/map/components/icon-tooltips/components/camera-info-popup.tsx
@@ -2,8 +2,8 @@
 import { useEffect, useState } from 'react'
 
 import { TooltipInfoItem } from '@/app/(app)/mapa/components/common/tooltip-info-item'
+import { Spinner } from '@/components/custom/spinner'
 import { Card } from '@/components/ui/card'
-import { Spinner } from '@/components/ui/spinner'
 import { useMap } from '@/hooks/use-contexts/use-map-context'
 import { cn } from '@/lib/utils'
