+++ src/app/(app)/mapa/components/map/components/map-actions.tsx
@@ -1,11 +1,11 @@
 import { TriangleAlert } from 'lucide-react'
 import type { Dispatch, SetStateAction } from 'react'
 
+import { Spinner } from '@/components/custom/spinner'
+import { Tooltip } from '@/components/custom/tooltip'
 import { Card } from '@/components/ui/card'
 import { Label } from '@/components/ui/label'
-import { Spinner } from '@/components/ui/spinner'
 import { Switch } from '@/components/ui/switch'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useMap } from '@/hooks/use-contexts/use-map-context'
 import { cn } from '@/lib/utils'
