+++ src/app/(app)/mapa/components/side-pannel/components/search-by-plate/components/action-buttons/components/monitoring-toggle.tsx
@@ -4,8 +4,8 @@ import { Siren } from 'lucide-react'
 import { useEffect, useState } from 'react'
 
 import { MonitoredPlateFormDialog } from '@/app/(app)/components/monitored-plate-form-dialog'
+import { Tooltip } from '@/components/custom/tooltip'
 import { Toggle } from '@/components/ui/toggle'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useMap } from '@/hooks/use-contexts/use-map-context'
 import { useMonitoredPlates } from '@/hooks/use-contexts/use-monitored-plates-context'
 import { useDisclosure } from '@/hooks/use-disclosure'
