+++ src/app/(app)/mapa/components/side-pannel/components/search-by-plate/components/trip-list/components/point-card.tsx
@@ -1,8 +1,8 @@
 import { formatDate } from 'date-fns'
 import { TriangleAlert } from 'lucide-react'
 
+import { Tooltip } from '@/components/custom/tooltip'
 import { Card, CardContent, CardHeader, CardTitle } from '@/components/ui/card'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useMap } from '@/hooks/use-contexts/use-map-context'
 import type { Point } from '@/models/entities'
 import { haversineDistance } from '@/utils/haversine-distance'
