+++ src/app/(app)/ocorrencias/components/content/components/side-panel/index.tsx
@@ -4,9 +4,9 @@ import { usePathname, useRouter } from 'next/navigation'
 import { Controller, useForm } from 'react-hook-form'
 import { z } from 'zod'
 
+import { DatePickerWithRange } from '@/components/custom/date-range-picker'
 import { MultiselectCheckbox } from '@/components/custom/multiselect-checkbox'
 import { Button } from '@/components/ui/button'
-import { DatePickerWithRange } from '@/components/ui/date-range-picker'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
 import { useReportsSearchParams } from '@/hooks/use-params/use-reports-search-params'
