+++ src/app/(app)/operacoes/components/operation-dialogs/components/operation-form-dialog.tsx
@@ -5,6 +5,8 @@ import { useEffect, useState } from 'react'
 import { useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 
+import { InputError } from '@/components/custom/input-error'
+import { Spinner } from '@/components/custom/spinner'
 import { Button } from '@/components/ui/button'
 import {
   Dialog,
@@ -13,9 +15,7 @@ import {
   DialogTitle,
 } from '@/components/ui/dialog'
 import { Input } from '@/components/ui/input'
-import { InputError } from '@/components/ui/input-error'
 import { Label } from '@/components/ui/label'
-import { Spinner } from '@/components/ui/spinner'
 import { Textarea } from '@/components/ui/textarea'
 import {
   type OperationForm,
