+++ src/app/(app)/operacoes/components/operations-header/index.tsx
@@ -1,6 +1,6 @@
 'use client'
+import { Tooltip } from '@/components/custom/tooltip'
 import { Button } from '@/components/ui/button'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useOperations } from '@/hooks/use-contexts/use-operations-context'
 import { useProfile } from '@/hooks/use-queries/use-profile'
 import { notAllowed } from '@/utils/template-messages'
