+++ src/app/(app)/operacoes/components/operations-table/index.tsx
@@ -3,10 +3,10 @@ import { useQuery } from '@tanstack/react-query'
 import { type ColumnDef } from '@tanstack/react-table'
 import { PencilLine, Trash } from 'lucide-react'
 
+import { Tooltip } from '@/components/custom/tooltip'
 import { Button } from '@/components/ui/button'
 import { DataTable } from '@/components/ui/data-table'
 import { Pagination } from '@/components/ui/pagination'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useOperations } from '@/hooks/use-contexts/use-operations-context'
 import { useOperationsSearchParams } from '@/hooks/use-params/use-operations-search-params'
 import { useProfile } from '@/hooks/use-queries/use-profile'
