+++ src/app/(app)/placas-monitoradas/components/monitored-plates-filter/index.tsx
@@ -6,6 +6,7 @@ import { useEffect } from 'react'
 import { Controller, useForm } from 'react-hook-form'
 import { z } from 'zod'
 
+import { Tooltip } from '@/components/custom/tooltip'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
@@ -16,7 +17,6 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select'
-import { Tooltip } from '@/components/ui/tooltip'
 
 const activeOptions = ['all', 'true', 'false']
