+++ src/app/(app)/placas-monitoradas/components/monitored-plates-header/index.tsx
@@ -1,7 +1,7 @@
 'use client'
 
+import { Tooltip } from '@/components/custom/tooltip'
 import { Button } from '@/components/ui/button'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useMonitoredPlates } from '@/hooks/use-contexts/use-monitored-plates-context'
 import { useProfile } from '@/hooks/use-queries/use-profile'
 import { notAllowed } from '@/utils/template-messages'
