+++ src/app/(app)/placas-monitoradas/components/monitored-plates-table/index.tsx
@@ -4,11 +4,11 @@ import { type ColumnDef } from '@tanstack/react-table'
 import { PencilLine, Trash } from 'lucide-react'
 import { useState } from 'react'
 
+import { Tooltip } from '@/components/custom/tooltip'
 import { Button } from '@/components/ui/button'
 import { DataTable } from '@/components/ui/data-table'
 import { Pagination } from '@/components/ui/pagination'
 import { Switch } from '@/components/ui/switch'
-import { Tooltip } from '@/components/ui/tooltip'
 import { useMonitoredPlates } from '@/hooks/use-contexts/use-monitored-plates-context'
 import { useMonitoredPlatesSearchParams } from '@/hooks/use-params/use-monitored-plates-search-params'
 import { useProfile } from '@/hooks/use-queries/use-profile'
