+++ src/app/auth/sign-in/sign-in-form.tsx
@@ -3,11 +3,11 @@ import { AlertTriangle, icons } from 'lucide-react'
 import { redirect } from 'next/navigation'
 import { useState } from 'react'
 
+import { Spinner } from '@/components/custom/spinner'
 import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
-import { Spinner } from '@/components/ui/spinner'
 import { useFormState } from '@/hooks/use-form-state'
 
 import { signInAction } from './actions'
