+++ src/components/custom/multiselect-checkbox.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react'
 import { Button } from '../ui/button'
 import { Checkbox } from '../ui/checkbox'
 import { Popover, PopoverContent, PopoverTrigger } from '../ui/popover'
-import { Spinner } from '../ui/spinner'
+import { Spinner } from './spinner'
 
 interface MultiselectCheckboxProps {
   value: string[] | undefined
