+++ src/components/custom/select-with-search.tsx
@@ -3,10 +3,7 @@
 import { CheckIcon, ChevronsUpDown } from 'lucide-react'
 import { useState } from 'react'
 
-import { cn } from '@/lib/utils'
-import type { ComboboxOption } from '@/models/utils'
-
-import { Button } from './button'
+import { Button } from '@/components/ui/button'
 import {
   Command,
   CommandEmpty,
@@ -14,8 +11,14 @@ import {
   CommandInput,
   CommandItem,
   CommandList,
-} from './command'
-import { Popover, PopoverContent, PopoverTrigger } from './popover'
+} from '@/components/ui/command'
+import {
+  Popover,
+  PopoverContent,
+  PopoverTrigger,
+} from '@/components/ui/popover'
+import { cn } from '@/lib/utils'
+import type { ComboboxOption } from '@/models/utils'
 
 interface SelectWithSearchProps {
   value: string
