+++ src/components/custom/time-picker.tsx
@@ -6,7 +6,7 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from './select'
+} from '../ui/select'
 
 const hours = Array.from({ length: 24 }, (_, i) =>
   i.toString().padStart(2, '0'),
